store/wizard: extract registration types into own submodule
Follows the pattern established by the other subslices (cloud, output,
etc.) — the registration type definitions now live in their own
directory so selectors, state, and slice logic can follow.
