+++ src/store/slices/wizard/index.ts
@@ -8,3 +8,4 @@ export * from './compliance/';
 export * from './details';
 export * from './filesystem';
 export * from './output';
+export * from './registration';
