+++ src/store/slices/wizard/registration/index.ts
@@ -1,0 +1,1 @@
+export * from './types';
